// packages/diagram/src/LikeC4Diagram.props.d.ts
type Any = t.aux.Any;
type Unknown = t.aux.UnknownLayouted;
type ViewId<A> = t.aux.ViewId<A>;
type Fqn<A> = t.aux.Fqn<A>;
type RelationId = t.aux.RelationId;
type DeploymentFqn<A> = t.aux.DeploymentFqn<A>;
type StrictViewId<A> = t.aux.StrictViewId<A>;
// ...
export type OpenSourceParams<A extends Any = Unknown> = {
    element: Fqn<A>;
    property?: string;
} | {
    relation: RelationId;
} | {
    deployment: DeploymentFqn<A>;
    property?: string;
} | {
    view: StrictViewId<A>;
} | {
    view: StrictViewId<A>;
    astPath: string;
};