// packages/diagram/src/likec4diagram/types.d.ts
export declare namespace Types {
// ...
    type ElementNodeData = Simplify<LeafNodeData & {
        modelFqn: Fqn;
        deploymentFqn?: never;
        /**
         * If set - this node has navigation to another view and diagram has handler for this
         */
        navigateTo: ViewId | null;
    }>;
    /**
     * Represents element from deployment model
     */
    type DeploymentElementNodeData = Simplify<LeafNodeData & {
        navigateTo: ViewId | null;
        deploymentFqn: DeploymentFqn;
        modelFqn: Fqn | null;
    }>;
    type SequenceActorNodePort = {
        id: string;
        cx: number;
        cy: number;
        height: number;
        type: 'target' | 'source';
        position: 'left' | 'right' | 'top' | 'bottom';
    };
    type SequenceActorNodeData = Simplify<LeafNodeData & {
        navigateTo: ViewId | null;
        modelFqn: Fqn | null;
        ports: Array<SequenceActorNodePort>;
        viewHeight: number;
    }>;
    type SequenceParallelAreaData = Simplify<LeafNodeData & {
        parallelPrefix: string;
    }>;
// ...
    type CompoundElementNodeData = Simplify<CompoundNodeData & {
        modelFqn: Fqn;
        deploymentFqn?: never;
        /**
         * If set - this node has navigation to another view and diagram has handler for this
         */
        navigateTo: ViewId | null;
    }>;
    type CompoundDeploymentNodeData = Simplify<CompoundNodeData & {
        deploymentFqn: DeploymentFqn;
        /**
         * If set - this node refers to a model element
         */
        modelFqn: Fqn | null;
        /**
         * If set - this node has navigation to another view and diagram has handler for this
         */
        navigateTo: ViewId | null;
    }>;
    type ViewGroupNodeData = Simplify<CompoundNodeData & {
        isViewGroup: true;
    }>;
// ...
    type NodeData = ExclusiveUnion<{
        ElementNodeData: ElementNodeData;
        DeploymentElementNodeData: DeploymentElementNodeData;
        CompoundElementNodeData: CompoundElementNodeData;
        CompoundDeploymentNodeData: CompoundDeploymentNodeData;
        ViewGroupNodeData: ViewGroupNodeData;
        SequenceActorNodeData: SequenceActorNodeData;
        SequenceParallelAreaData: SequenceParallelAreaData;
    }>;
    type Node<Type extends NodeType = NodeType> = Extract<AnyNode, {
        type: Type;
    }>;
    type NodeProps<Type extends NodeType = NodeType> = BaseNodeProps<Node<Type>>;
    type RelationshipEdgeData = Simplify<BaseEdgeData & NonOptional<Pick<DiagramEdge, 'id' | 'label' | 'labelBBox' | 'technology' | 'points' | 'dir' | 'color' | 'line' | 'head' | 'tail' | 'navigateTo' | 'astPath' | 'drifts'>> & {
        notes: MarkdownOrString | null;
        labelXY: XYPosition | null;
        controlPoints: XYPosition[] | undefined | null;
    }>;
    type SequenceStepEdgeData = Simplify<BaseEdgeData & NonOptional<Pick<DiagramEdge, 'id' | 'label' | 'technology' | 'points' | 'dir' | 'color' | 'line' | 'head' | 'tail' | 'navigateTo' | 'astPath' | 'drifts'>> & {
        notes: MarkdownOrString | null;
        labelXY: XYPosition | null;
        labelBBox: BBox;
        controlPoints: XYPosition[] | undefined | null;
    }>;
    type RelationshipEdge = BaseEdge<RelationshipEdgeData, 'relationship'>;
    type SequenceStepEdge = BaseEdge<SequenceStepEdgeData, 'seq-step'>;
    type AnyEdge = RelationshipEdge | SequenceStepEdge;
    type EdgeType = AnyEdge['type'];
    type Edge<Type extends EdgeType = EdgeType> = Extract<AnyEdge, {
        type: Type;
    }>;
    type EdgeProps<Type extends EdgeType = EdgeType> = BaseEdgeProps<Edge<Type>>;
    type EdgeData = ExclusiveUnion<{
        RelationshipEdgeData: RelationshipEdgeData;
        SequenceStepEdgeData: SequenceStepEdgeData;
    }>;
    type NodeRenderer<T extends NodeType> = FunctionComponent<NodeProps<T>>;
    type NodeRenderers = {
        [T in NodeType]: NodeRenderer<T>;
    };
    type EdgeRenderer<T extends EdgeType> = FunctionComponent<EdgeProps<T>>;
    type EdgeRenderers = {
        [T in EdgeType]: EdgeRenderer<T>;
    };
}